Changed the placement of exception_handlers.py and renamed it to handlers.py in the exceptions folder. Deleted the api folder.
